move getExpandAdvancedPaneAction to actions.ts & default for showGuides should be true
