+++ ui/user/src/lib/services/guides/mcp/customMcp.ts
@@ -1,5 +1,5 @@
+import { getExpandAdvancedPaneAction } from '../actions';
 import type { GuideStep } from '../types';
-import { getExpandAdvancedPaneAction } from '../utils';
 import {
 	highlightMcpCatalogLink,
 	listenMcpCatalogLink,
